Add new extensions git and ssh
git mounts the Git config from the host into the container.
ssh mounts the host user's SSH config into the container and forwards
the SSH agent from the host.

Both together allow to work with Git and push and pull in the same way
as on the host, given that the respective packages are installed in the
container.